+++ lib/csvec.ml
@@ -42,6 +42,18 @@ let new_from_unsorted n indices data =
   new_checked n indices data
 
 let empty n = new_trusted n [||] [||]
+
+let of_dense ?(epsilon = 0.00001) a =
+  let indices = Dynarray.create () in
+  let data = Dynarray.create () in
+  Array.iteri
+    (fun i x ->
+      if abs_float x >= epsilon then (
+        Dynarray.add_last indices i;
+        Dynarray.add_last data x))
+    a;
+  { dim = Array.length a; indices; data }
+
 let nnz (v : 'a t) = Dynarray.length v.data
 
 let nnz_index (v : 'a t) index =
@@ -67,7 +79,7 @@ let fold f (acc : 'acc) (v : 'a t) =
 
 let iter f (v : 'a t) =
   let open Dynarray in
-  for i = 0 to length v.indices do
+  for i = 0 to length v.indices - 1 do
     f v.indices.!(i) v.data.!(i)
   done
 
@@ -84,6 +96,7 @@ let count f (v : 'a t) =
   !c
 
 let is_empty (v : 'a t) = Dynarray.is_empty v.data
+let scale c v = map (fun x -> x *. c) v
 
 let append (v : 'a t) ind x =
   let open Dynarray in